+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
-## Unreleased
-[Check the diff](https://github.com/elastic/apm-agent-python/compare/v5.1.0...master)
+## v5.1.1 
+[Check the diff](https://github.com/elastic/apm-agent-python/compare/v5.1.0...v5.1.1)
 
 ### Bugfixes
  * fixed an issue with empty responses from APM Server's config endpoint (#562, #563)
  * fixed Windows tests by avoiding time.sleep in breakdown metrics tests (#537, #550)
  * fixed container ID matching to match CloudFoundry Garden container IDs (#523, #564)
+ * fixed an issue in the urllib instrumentation if no port is set (#567)
+
+### Other
+ * Added Python 3.8 RC to the test matrix (#565)
 
 ## v5.1.0 
 [Check the diff](https://github.com/elastic/apm-agent-python/compare/v5.0.0...v5.1.0)